be/entity-update into backend-dev

See merge request s12-final/S12P31B103!6

+++ server/src/main/java/com/ssafy/odab/domain/concept/repository/SubConceptRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.odab.domain.concept.repository;
 import com.ssafy.odab.domain.concept.entity.SubConcept;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubConceptRepository extends JpaRepository<SubConcept, Long> {
+public interface SubConceptRepository extends JpaRepository<SubConcept, Integer> {
 
 }
